Many many changes, hope that i didnt break something
DONE:
Incorporated bootstrap to the design of the website for every page except from the User page.
Fixed profile page
Fixed a problem with the header where the user accidentally could click on the logo

TODO:
Create accept and decline invitation cards.
Create information windows when following a book
Refactor the .css files